requirements.txt
pandas
numpy
scikit-learn
ta
requests
beautifulsoup4
matplotlib
yfinance
schedule
python-telegram-bot
scipy
seaborn
plotly
xgboost
lightgbm
joblib
flask
pyyaml
gunicorn